chore: rowTitle로 지정된 컬럼이 가장 앞으로 오도록 설정
- 데이터에 들어갈 수 있는 타입 중 boolean 타입은 제거
- 인터페이스 파일 분리
